Trait sos21_domain::context::object_repository::ObjectRepository

pub trait ObjectRepository {
    type OutOfLimitSizeError: Error;

    // Required methods
    fn store_object<'life0, 'async_trait>(
        &'life0 self,
        object: Object
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn store_object_with_limit<'life0, 'async_trait>(
        &'life0 self,
        object: Object,
        limit: u64
    ) -> Pin<Box<dyn Future<Output = Result<Result<(), Self::OutOfLimitSizeError>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_object<'life0, 'async_trait>(
        &'life0 self,
        id: ObjectId
    ) -> Pin<Box<dyn Future<Output = Result<Option<Object>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Required Associated Types

type OutOfLimitSizeError: Error

Required Methods

fn store_object<'life0, 'async_trait>( &'life0 self, object: Object ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Stores an object.

fn store_object_with_limit<'life0, 'async_trait>( &'life0 self, object: Object, limit: u64 ) -> Pin<Box<dyn Future<Output = Result<Result<(), Self::OutOfLimitSizeError>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Stores an object while specifying the limit of object size.

When the size of the object is larger than the limit ( size > limit ), the object will not be stored and Self::OutOfLimitSizeError is returned.

fn get_object<'life0, 'async_trait>( &'life0 self, id: ObjectId ) -> Pin<Box<dyn Future<Output = Result<Option<Object>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,