Struct sos21_domain::model::form::Form

pub struct Form {
    content: FormContent,
}

Fields
content: FormContent
Implementations

impl Form

pub fn new(
author: &User,
name: FormName,
description: FormDescription,
period: FormPeriod,
items: FormItems,
condition: FormCondition
) -> Result<Self, NewFormError>

pub async fn answer_by<C>(
&self,
ctx: C,
user: &User,
project: &Project,
items: FormAnswerItems
) -> DomainResult<FormAnswer, AnswerError> where
C: FormAnswerRepository,

pub fn can_be_answered_by(&self, user: &User, project: &Project) -> bool

pub fn from_content(content: FormContent) -> Self

Restore Form from FormContent.

This is intended to be used when the data is taken out of the implementation by Form::into_content for persistence, internal serialization, etc. Use Form::new to create a form.

pub fn into_content(self) -> FormContent

Convert Form into FormContent.

pub fn id(&self) -> FormId

pub fn created_at(&self) -> DateTime

pub fn author_id(&self) -> &UserId

pub fn name(&self) -> &FormName

pub fn description(&self) -> &FormDescription

pub fn period(&self) -> FormPeriod

pub fn items(&self) -> &FormItems

pub fn condition(&self) -> &FormCondition

pub fn into_items(self) -> FormItems

impl Form

fn require_update_permission(
&self,
user: &User
) -> Result<(), NoUpdatePermissionError>

pub fn set_name(
&mut self,
user: &User,
name: FormName
) -> Result<(), NoUpdatePermissionError>

pub fn set_description(
&mut self,
user: &User,
description: FormDescription
) -> Result<(), NoUpdatePermissionError>

pub fn set_period(
&mut self,
user: &User,
period: FormPeriod
) -> Result<(), SetPeriodError>

pub fn set_items(
&mut self,
user: &User,
items: FormItems
) -> Result<(), NoUpdatePermissionError>

pub fn set_condition(
&mut self,
user: &User,
condition: FormCondition
) -> Result<(), NoUpdatePermissionError>

pub fn is_visible_to(&self, user: &User) -> bool

pub fn is_visible_to_with_project(&self, user: &User, project: &Project) -> bool