Struct sos21_domain::model::user::User

pub struct User {
    content: UserContent,
}

Fields
content: UserContent
Implementations

impl User

pub async fn new<C>(
ctx: &C,
id: UserId,
name: UserName,
kana_name: UserKanaName,
phone_number: PhoneNumber,
affiliation: UserAffiliation,
email: UserEmailAddress,
category: UserCategory
) -> DomainResult<Self, AlreadySignedUpError> where
C: UserRepository + UserInvitationRepository,

pub fn from_content(content: UserContent) -> Self

Restore User from UserContent.

This is intended to be used when the data is taken out of the implementation by User::into_content for persistence, internal serialization, etc. Use User::new to create a project.

pub fn into_content(self) -> UserContent

Convert User into UserContent.

pub fn id(&self) -> &UserId

pub fn created_at(&self) -> DateTime

pub fn name(&self) -> &UserName

pub fn kana_name(&self) -> &UserKanaName

pub fn phone_number(&self) -> &PhoneNumber

pub fn affiliation(&self) -> &UserAffiliation

pub fn email(&self) -> &UserEmailAddress

pub fn role(&self) -> UserRole

pub fn category(&self) -> UserCategory

pub fn assignment(&self) -> Option<UserAssignment>

pub fn assign_project_owner(&mut self, project: &Project) -> Result<()>

pub fn assign_project_subowner(&mut self, project: &Project) -> Result<()>

pub fn assign_pending_project_owner(
&mut self,
pending_project: &PendingProject
) -> Result<()>

pub fn permissions(&self) -> Permissions

pub fn require_permissions(
&self,
permissions: Permissions
) -> Result<(), RequirePermissionsError>

pub fn is_visible_to(&self, user: &User) -> bool

pub async fn file_usage<C>(&self, ctx: C) -> Result<UserFileUsage> where
C: FileRepository,

pub fn file_usage_quota(&self) -> UserFileUsageQuota

pub fn is_committee(&self) -> bool

pub fn is_committee_operator(&self) -> bool

pub fn set_name(&mut self, name: UserName)

pub fn set_kana_name(&mut self, kana_name: UserKanaName)

pub fn set_phone_number(&mut self, phone_number: PhoneNumber)

pub fn set_affiliation(&mut self, affiliation: UserAffiliation)

pub fn set_role(&mut self, role: UserRole)

pub fn set_category(&mut self, category: UserCategory)