Struct sos21_domain::test::context::MockApp

pub struct MockApp {
    users: Arc<Mutex<HashMap<UserId, User>>>,
    projects: Arc<Mutex<HashMap<ProjectId, Project>>>,
    forms: Arc<Mutex<HashMap<FormId, Form>>>,
    answers: Arc<Mutex<HashMap<FormAnswerId, FormAnswer>>>,
    files: Arc<Mutex<HashMap<FileId, File>>>,
    objects: Arc<Mutex<HashMap<ObjectId, Bytes>>>,
    sharings: Arc<Mutex<HashMap<FileSharingId, FileSharing>>>,
    distributions: Arc<Mutex<HashMap<FileDistributionId, FileDistribution>>>,
    pending_projects: Arc<Mutex<HashMap<PendingProjectId, PendingProject>>>,
    registration_forms: Arc<Mutex<HashMap<RegistrationFormId, RegistrationForm>>>,
    registration_form_answers: Arc<Mutex<HashMap<RegistrationFormAnswerId, RegistrationFormAnswer>>>,
    user_invitations: Arc<Mutex<HashMap<UserInvitationId, UserInvitation>>>,
}

Fields
users: Arc<Mutex<HashMap<UserId, User>>>projects: Arc<Mutex<HashMap<ProjectId, Project>>>forms: Arc<Mutex<HashMap<FormId, Form>>>answers: Arc<Mutex<HashMap<FormAnswerId, FormAnswer>>>files: Arc<Mutex<HashMap<FileId, File>>>objects: Arc<Mutex<HashMap<ObjectId, Bytes>>>sharings: Arc<Mutex<HashMap<FileSharingId, FileSharing>>>distributions: Arc<Mutex<HashMap<FileDistributionId, FileDistribution>>>pending_projects: Arc<Mutex<HashMap<PendingProjectId, PendingProject>>>registration_forms: Arc<Mutex<HashMap<RegistrationFormId, RegistrationForm>>>registration_form_answers: Arc<Mutex<HashMap<RegistrationFormAnswerId, RegistrationFormAnswer>>>user_invitations: Arc<Mutex<HashMap<UserInvitationId, UserInvitation>>>
Implementations

impl MockApp

pub fn authenticate_as(
self,
user_id: String,
email: String
) -> Authentication<MockApp>

Panics

This function panics when the given email is not valid.

pub async fn login_as(self, user: User) -> Login<MockApp>

Panics

This function panics when the login is not successful.