Trait sos21_domain::context::object_repository::ObjectRepository

pub trait ObjectRepository {
    type OutOfLimitSizeError: Error;
#[must_use]    fn store_object<'life0, 'async_trait>(
        &'life0 self, 
        object: Object
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn store_object_with_limit<'life0, 'async_trait>(
        &'life0 self, 
        object: Object, 
        limit: u64
    ) -> Pin<Box<dyn Future<Output = Result<Result<(), Self::OutOfLimitSizeError>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn get_object<'life0, 'async_trait>(
        &'life0 self, 
        id: ObjectId
    ) -> Pin<Box<dyn Future<Output = Result<Option<Object>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Associated Types

type OutOfLimitSizeError: Error

Loading content...
Required methods

#[must_use]fn store_object<'life0, 'async_trait>(
&'life0 self,
object: Object
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Stores an object.

#[must_use]fn store_object_with_limit<'life0, 'async_trait>(
&'life0 self,
object: Object,
limit: u64
) -> Pin<Box<dyn Future<Output = Result<Result<(), Self::OutOfLimitSizeError>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Stores an object while specifying the limit of object size.

When the size of the object is larger than the limit ( size > limit ), the object will not be stored and Self::OutOfLimitSizeError is returned.

#[must_use]fn get_object<'life0, 'async_trait>(
&'life0 self,
id: ObjectId
) -> Pin<Box<dyn Future<Output = Result<Option<Object>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Loading content...