Struct sos21_domain::model::project::Project

pub struct Project {
    content: ProjectContent,
    owner_id: UserId,
    subowner_id: UserId,
}

Fields
content: ProjectContentowner_id: UserIdsubowner_id: UserId
Implementations

impl Project

pub async fn new<C>(
ctx: C,
pending_project: PendingProject,
subowner: &User
) -> Result<Result<Self, NewProjectError>> where
C: ProjectRepository + RegistrationFormRepository + RegistrationFormAnswerRepository,

pub fn from_content(
content: ProjectContent,
owner_id: UserId,
subowner_id: UserId
) -> Result<Self, SameOwnerSubownerError>

Restore Project from ProjectContent.

This is intended to be used when the data is taken out of the implementation by Project::into_content for persistence, internal serialization, etc. Use Project::new to create a project.

pub fn into_content(self) -> ProjectContent

Convert Project into ProjectContent.

pub fn id(&self) -> ProjectId

pub fn index(&self) -> ProjectIndex

pub fn created_at(&self) -> DateTime

pub fn name(&self) -> &ProjectName

pub fn kana_name(&self) -> &ProjectKanaName

pub fn group_name(&self) -> &ProjectGroupName

pub fn kana_group_name(&self) -> &ProjectKanaGroupName

pub fn description(&self) -> &ProjectDescription

pub fn category(&self) -> ProjectCategory

pub fn attributes(&self) -> &ProjectAttributes

pub fn owner_id(&self) -> &UserId

pub fn subowner_id(&self) -> &UserId

pub fn is_visible_to(&self, user: &User) -> bool

pub fn kind(&self) -> ProjectKind

pub fn code(&self) -> ProjectCode

pub fn set_name(&mut self, name: ProjectName)

pub fn set_kana_name(&mut self, kana_name: ProjectKanaName)

pub fn set_group_name(&mut self, group_name: ProjectGroupName)

pub fn set_kana_group_name(&mut self, kana_group_name: ProjectKanaGroupName)

pub fn set_description(&mut self, description: ProjectDescription)

pub fn set_category(&mut self, category: ProjectCategory)

pub fn set_attributes(&mut self, attributes: ProjectAttributes)